Refactor ShiftsScreen into separate components
