+++ guard_app/src/screen/ShiftsScreen.tsx
@@ -1,76 +1,31 @@
 import { createMaterialTopTabNavigator } from '@react-navigation/material-top-tabs';
 import { useFocusEffect } from '@react-navigation/native';
-import React, { useCallback, useMemo, useState } from 'react';
+import React, { useCallback, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import {
   ActivityIndicator,
   Alert,
   Dimensions,
   FlatList,
-  Modal,
-  Pressable,
   RefreshControl,
   StyleSheet,
   Text,
   TextInput,
-  TouchableOpacity,
   View,
 } from 'react-native';
 
 import { getMe } from '../api/auth';
 import { applyToShift, listShifts, myShifts, type ShiftDto } from '../api/shifts';
+import ShiftCard from '../components/card/ShiftCard';
+import CalendarView from '../components/calendar/CalendarView';
+import ShiftDetailsModal from '../components/modal/ShiftDetailsModal';
+import ViewToggle from '../components/toggle/ViewToggle';
 import { useAppTheme } from '../theme';
+import type { AllShift, AppliedShift, CompletedShift } from '../models/Shifts';
 
 import type { AppColors } from '../theme/colors';
 
 const { width } = Dimensions.get('window');
-
-type AppliedShift = {
-  id: string;
-  title: string;
-  company: string;
-  site: string;
-  rate: string;
-  date: string;
-  time: string;
-  status?: 'Pending' | 'Confirmed' | 'Rejected';
-};
-
-type CompletedShift = {
-  id: string;
-  title: string;
-  company: string;
-  site: string;
-  rate: string;
-  date: string;
-  time: string;
-  rated: boolean;
-  rating: number;
-};
-
-type AllShift = {
-  id: string;
-  title: string;
-  company: string;
-  site: string;
-  rate: string;
-  date: string;
-  time: string;
-  status?: 'Available' | 'Pending' | 'Confirmed';
-};
-
-function dateKeyLocal(d: Date) {
-  const y = d.getFullYear();
-  const m = String(d.getMonth() + 1).padStart(2, '0');
-  const day = String(d.getDate()).padStart(2, '0');
-  return `${y}-${m}-${day}`;
-}
-
-function addDays(d: Date, days: number) {
-  const copy = new Date(d);
-  copy.setDate(copy.getDate() + days);
-  return copy;
-}
 
 function mapMineShifts(shifts: ShiftDto[], myUid: string): AppliedShift[] {
   return shifts
@@ -145,333 +100,6 @@ function mapAllShifts(shifts: ShiftDto[], myUid: string): AllShift[] {
       };
     });
 }
-
-function ShiftDetailsModal({
-  shift,
-  visible,
-  onClose,
-  colors,
-}: {
-  shift: AppliedShift | CompletedShift | AllShift | null;
-  visible: boolean;
-  onClose: () => void;
-  colors: AppColors;
-}) {
-  const s = getStyles(colors);
-  const { t } = useTranslation();
-
-  if (!shift) return null;
-
-  const status = 'status' in shift ? shift.status : 'Completed';
-  const statusColor =
-    status === 'Confirmed'
-      ? colors.status.confirmed
-      : status === 'Pending'
-        ? colors.link
-        : status === 'Available'
-          ? colors.primary
-          : colors.muted;
-
-  return (
-    <Modal visible={visible} transparent animationType="fade" onRequestClose={onClose}>
-      <Pressable style={s.modalOverlay} onPress={onClose}>
-        <Pressable style={s.modalContent} onPress={(e) => e.stopPropagation()}>
-          <View style={s.modalHeader}>
-            <Text style={s.modalTitle}>{t('shifts.shiftDetails')}</Text>
-            <TouchableOpacity onPress={onClose} style={s.modalCloseBtn}>
-              <Text style={s.modalCloseText}>✕</Text>
-            </TouchableOpacity>
-          </View>
-
-          <View style={s.modalBody}>
-            <View style={s.modalTitleRow}>
-              <Text style={s.modalShiftTitle}>{shift.title}</Text>
-              <View style={[s.statusBadge, { backgroundColor: statusColor }]}>
-                <Text style={s.statusBadgeText}>
-                  {status ? t(`shifts.${status.toLowerCase()}`, status) : t('shifts.available')}
-                </Text>
-              </View>
-            </View>
-
-            <View style={s.modalDetail}>
-              <Text style={s.modalLabel}>{t('shifts.date')}</Text>
-              <Text style={s.modalValue}>
-                {new Date(shift.date).toLocaleDateString('en-US', {
-                  weekday: 'short',
-                  month: 'short',
-                  day: 'numeric',
-                  year: 'numeric',
-                })}
-              </Text>
-            </View>
-
-            <View style={s.modalDetail}>
-              <Text style={s.modalLabel}>{t('shifts.time')}</Text>
-              <Text style={s.modalValue}>{shift.time}</Text>
-            </View>
-
-            <View style={s.modalDetail}>
-              <Text style={s.modalLabel}>{t('shifts.location')}</Text>
-              <Text style={s.modalValue}>{shift.site}</Text>
-            </View>
-
-            <View style={s.modalDetail}>
-              <Text style={s.modalLabel}>{t('shifts.payRate')}</Text>
-              <Text style={s.modalValue}>{shift.rate}</Text>
-            </View>
-
-            <View style={s.modalRequirements}>
-              <Text style={s.modalRequirementsTitle}>{t('shifts.requirements')}</Text>
-              <View style={s.modalTags}>
-                <View style={s.modalTag}>
-                  <Text style={s.modalTagText}>Security License</Text>
-                </View>
-                <View style={s.modalTag}>
-                  <Text style={s.modalTagText}>First Aid</Text>
-                </View>
-              </View>
-            </View>
-          </View>
-        </Pressable>
-      </Pressable>
-    </Modal>
-  );
-}
-
-function CalendarView<T extends { id: string; date: string; title: string; status?: string }>({
-  shifts,
-  onShiftPress,
-  colors,
-}: {
-  shifts: T[];
-  onShiftPress: (shift: T) => void;
-  colors: AppColors;
-}) {
-  const s = getStyles(colors);
-  const { t } = useTranslation();
-  const [monthCursor, setMonthCursor] = useState(() => new Date());
-
-  const monthStart = new Date(monthCursor.getFullYear(), monthCursor.getMonth(), 1);
-  const monthEnd = new Date(monthCursor.getFullYear(), monthCursor.getMonth() + 1, 0);
-  const startWeekday = monthStart.getDay();
-  const gridStart = addDays(monthStart, -startWeekday);
-
-  const totalCells = 42;
-  const allCells = Array.from({ length: totalCells }, (_, i) => addDays(gridStart, i));
-
-  const shiftsByDate = useMemo(() => {
-    const map = new Map<string, T[]>();
-    shifts.forEach((shift) => {
-      const key = dateKeyLocal(new Date(shift.date));
-      if (!map.has(key)) map.set(key, []);
-      map.get(key)!.push(shift);
-    });
-    return map;
-  }, [shifts]);
-
-  const monthLabel = monthStart.toLocaleString(undefined, { month: 'long', year: 'numeric' });
-
-  const getStatusColor = (status?: string) => {
-    if (status === 'Confirmed') return colors.status.confirmed;
-    if (status === 'Pending') return colors.link;
-    if (status === 'Available') return colors.primary;
-    return colors.muted;
-  };
-
-  return (
-    <View style={s.calendarContainer}>
-      <View style={s.calHeader}>
-        <Text style={s.calMonthText}>{monthLabel}</Text>
-        <View style={s.calNavButtons}>
-          <TouchableOpacity
-            onPress={() =>
-              setMonthCursor(new Date(monthCursor.getFullYear(), monthCursor.getMonth() - 1, 1))
-            }
-            style={s.calNavBtn}
-          >
-            <Text style={s.calNavBtnText}>‹</Text>
-          </TouchableOpacity>
-          <TouchableOpacity
-            onPress={() =>
-              setMonthCursor(new Date(monthCursor.getFullYear(), monthCursor.getMonth() + 1, 1))
-            }
-            style={s.calNavBtn}
-          >
-            <Text style={s.calNavBtnText}>›</Text>
-          </TouchableOpacity>
-        </View>
-      </View>
-
-      <View style={s.calWeekHeader}>
-        {['Sun', 'Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat'].map((day, i) => (
-          <View key={i} style={s.calWeekCell}>
-            <Text style={s.calWeekText}>{day}</Text>
-          </View>
-        ))}
-      </View>
-
-      <View style={s.calGrid}>
-        {allCells.map((d, idx) => {
-          const inMonth = d >= monthStart && d <= monthEnd;
-          const key = dateKeyLocal(d);
-          const dayShifts = shiftsByDate.get(key) || [];
-          const hasShifts = dayShifts.length > 0;
-
-          return (
-            <TouchableOpacity
-              key={idx}
-              style={[s.calDayCell, !inMonth && s.calDayCellDim]}
-              onPress={() => hasShifts && onShiftPress(dayShifts[0])}
-              disabled={!hasShifts}
-            >
-              <Text style={[s.calDayNumber, !inMonth && s.calDayNumberDim]}>{d.getDate()}</Text>
-              {hasShifts && (
-                <View style={s.calShiftIndicators}>
-                  {dayShifts.slice(0, 3).map((shift, i) => (
-                    <View
-                      key={i}
-                      style={[s.calShiftDot, { backgroundColor: getStatusColor(shift.status) }]}
-                    />
-                  ))}
-                </View>
-              )}
-            </TouchableOpacity>
-          );
-        })}
-      </View>
-
-      <View style={s.calLegend}>
-        <View style={s.calLegendItem}>
-          <View style={[s.calLegendDot, { backgroundColor: colors.primary }]} />
-          <Text style={s.calLegendText}>{t('shifts.available')}</Text>
-        </View>
-        <View style={s.calLegendItem}>
-          <View style={[s.calLegendDot, { backgroundColor: colors.link }]} />
-          <Text style={s.calLegendText}>{t('shifts.applied')}</Text>
-        </View>
-        <View style={s.calLegendItem}>
-          <View style={[s.calLegendDot, { backgroundColor: colors.status.confirmed }]} />
-          <Text style={s.calLegendText}>{t('shifts.accepted')}</Text>
-        </View>
-        <View style={s.calLegendItem}>
-          <View style={[s.calLegendDot, { backgroundColor: colors.muted }]} />
-          <Text style={s.calLegendText}>{t('shifts.completed')}</Text>
-        </View>
-      </View>
-    </View>
-  );
-}
-
-function ShiftCard({
-  shift,
-  onPress,
-  colors,
-  showApply = false,
-  onApply,
-  applying = false,
-}: {
-  shift: AppliedShift | CompletedShift | AllShift;
-  onPress?: () => void;
-  colors: AppColors;
-  showApply?: boolean;
-  onApply?: () => void;
-  applying?: boolean;
-}) {
-  const s = getStyles(colors);
-  const { t } = useTranslation();
-
-  const status = 'status' in shift ? shift.status : 'Completed';
-
-  const statusColor =
-    status === 'Confirmed'
-      ? colors.status.confirmed
-      : status === 'Pending'
-        ? colors.link
-        : status === 'Available'
-          ? colors.primary
-          : colors.muted;
-
-  return (
-    <TouchableOpacity style={s.card} onPress={onPress} activeOpacity={0.9}>
-      <View style={s.cardHeader}>
-        <View style={s.cardTitleSection}>
-          <Text style={s.cardTitle}>{shift.title}</Text>
-
-          {showApply && status === 'Available' ? (
-            <TouchableOpacity
-              style={[s.applyBtn, applying && s.applyBtnDisabled]}
-              onPress={onApply}
-              disabled={applying}
-            >
-              <Text style={s.applyBtnText}>
-                {applying ? t('shifts.applying') : t('shifts.apply')}
-              </Text>
-            </TouchableOpacity>
-          ) : (
-            <View style={[s.cardStatusBadge, { backgroundColor: statusColor }]}>
-              <Text style={s.cardStatusText}>
-                {status ? t(`shifts.${status.toLowerCase()}`, status) : t('shifts.available')}
-              </Text>
-            </View>
-          )}
-        </View>
-      </View>
-
-      <Text style={s.cardCompany}>{shift.company}</Text>
-
-      <View style={s.cardRow}>
-        <Text style={s.cardLabel}>{t('shifts.date')}</Text>
-        <Text style={s.cardValue}>
-          {new Date(shift.date).toLocaleDateString('en-US', {
-            month: 'short',
-            day: 'numeric',
-            year: 'numeric',
-          })}
-        </Text>
-      </View>
-
-      <View style={s.cardRow}>
-        <Text style={s.cardLabel}>{t('shifts.time')}</Text>
-        <Text style={s.cardValue}>{shift.time}</Text>
-      </View>
-
-      <View style={s.cardRow}>
-        <Text style={s.cardLabel}>{t('shifts.payRate')}</Text>
-        <Text style={s.cardPay}>{shift.rate}</Text>
-      </View>
-    </TouchableOpacity>
-  );
-}
-
-function ViewToggle({
-  view,
-  onViewChange,
-  colors,
-}: {
-  view: 'list' | 'calendar';
-  onViewChange: (view: 'list' | 'calendar') => void;
-  colors: AppColors;
-}) {
-  const s = getStyles(colors);
-
-  return (
-    <View style={s.viewToggle}>
-      <TouchableOpacity
-        style={[s.viewToggleBtn, view === 'list' && s.viewToggleBtnActive]}
-        onPress={() => onViewChange('list')}
-      >
-        <Text style={[s.viewToggleIcon, view === 'list' && s.viewToggleIconActive]}>☰</Text>
-      </TouchableOpacity>
-      <TouchableOpacity
-        style={[s.viewToggleBtn, view === 'calendar' && s.viewToggleBtnActive]}
-        onPress={() => onViewChange('calendar')}
-      >
-        <Text style={[s.viewToggleIcon, view === 'calendar' && s.viewToggleIconActive]}>📅</Text>
-      </TouchableOpacity>
-    </View>
-  );
-}
-
 function AllTab() {
   const { colors } = useAppTheme();
   const s = getStyles(colors);
@@ -711,7 +339,7 @@ function CompletedTab() {
 
       {view === 'calendar' ? (
         <CalendarView
-          shifts={filtered.map((s) => ({ ...s, status: 'Completed' }))}
+          shifts={filtered.map((s) => ({ ...s, status: 'Completed' as const }))}
           onShiftPress={setSelectedShift}
           colors={colors}
         />
@@ -816,320 +444,6 @@ const getStyles = (colors: AppColors) =>
     searchInput: {
       flex: 1,
       fontSize: 14,
-      color: colors.text,
-    },
-
-    viewToggle: {
-      flexDirection: 'row',
-      backgroundColor: colors.card,
-      borderRadius: 8,
-      borderWidth: 1,
-      borderColor: colors.border,
-      overflow: 'hidden',
-    },
-    viewToggleBtn: {
-      width: 44,
-      height: 44,
-      alignItems: 'center',
-      justifyContent: 'center',
-    },
-    viewToggleBtnActive: {
-      backgroundColor: colors.primary,
-    },
-    viewToggleIcon: {
-      fontSize: 18,
-      color: colors.muted,
-    },
-    viewToggleIconActive: {
-      color: colors.white,
-    },
-
-    card: {
-      backgroundColor: colors.card,
-      borderRadius: 12,
-      padding: 16,
-      marginBottom: 12,
-      borderWidth: 1,
-      borderColor: colors.border,
-    },
-    cardHeader: {
-      marginBottom: 8,
-    },
-    cardTitleSection: {
-      flexDirection: 'row',
-      justifyContent: 'space-between',
-      alignItems: 'center',
-    },
-    cardTitle: {
-      fontSize: 16,
-      fontWeight: '700',
-      color: colors.text,
-      flex: 1,
-    },
-    cardStatusBadge: {
-      paddingHorizontal: 10,
-      paddingVertical: 4,
-      borderRadius: 12,
-      marginLeft: 12,
-    },
-    cardStatusText: {
-      fontSize: 11,
-      fontWeight: '600',
-      color: colors.white,
-    },
-    cardCompany: {
-      fontSize: 13,
-      color: colors.muted,
-      marginBottom: 12,
-    },
-    cardRow: {
-      flexDirection: 'row',
-      marginBottom: 6,
-    },
-    cardLabel: {
-      fontSize: 13,
-      color: colors.muted,
-      width: 60,
-    },
-    cardValue: {
-      fontSize: 13,
-      color: colors.text,
-      fontWeight: '500',
-    },
-    cardPay: {
-      fontSize: 13,
-      color: colors.status.confirmed,
-      fontWeight: '700',
-    },
-
-    applyBtn: {
-      backgroundColor: colors.primary,
-      paddingHorizontal: 14,
-      paddingVertical: 8,
-      borderRadius: 12,
-      marginLeft: 12,
-    },
-    applyBtnDisabled: {
-      opacity: 0.7,
-    },
-    applyBtnText: {
-      color: colors.white,
-      fontSize: 12,
-      fontWeight: '700',
-    },
-
-    calendarContainer: {
-      backgroundColor: colors.card,
-      borderRadius: 12,
-      padding: 16,
-      marginBottom: 16,
-      borderWidth: 1,
-      borderColor: colors.border,
-    },
-    calHeader: {
-      flexDirection: 'row',
-      justifyContent: 'space-between',
-      alignItems: 'center',
-      marginBottom: 16,
-    },
-    calMonthText: {
-      fontSize: 16,
-      fontWeight: '700',
-      color: colors.text,
-    },
-    calNavButtons: {
-      flexDirection: 'row',
-      gap: 8,
-    },
-    calNavBtn: {
-      width: 32,
-      height: 32,
-      backgroundColor: colors.primarySoft,
-      borderRadius: 6,
-      alignItems: 'center',
-      justifyContent: 'center',
-    },
-    calNavBtnText: {
-      fontSize: 18,
-      fontWeight: '600',
-      color: colors.text,
-    },
-    calWeekHeader: {
-      flexDirection: 'row',
-      marginBottom: 8,
-    },
-    calWeekCell: {
-      flex: 1,
-      alignItems: 'center',
-    },
-    calWeekText: {
-      fontSize: 12,
-      fontWeight: '600',
-      color: colors.muted,
-    },
-    calGrid: {
-      flexDirection: 'row',
-      flexWrap: 'wrap',
-      gap: 4,
-    },
-    calDayCell: {
-      width: (width - 32 - 16 * 2 - 24) / 7,
-      aspectRatio: 1,
-      borderRadius: 8,
-      borderWidth: 1,
-      borderColor: colors.border,
-      alignItems: 'center',
-      justifyContent: 'center',
-      padding: 4,
-      backgroundColor: colors.card,
-    },
-    calDayCellDim: {
-      opacity: 0.3,
-    },
-    calDayNumber: {
-      fontSize: 13,
-      fontWeight: '600',
-      color: colors.text,
-    },
-    calDayNumberDim: {
-      color: colors.muted,
-    },
-    calShiftIndicators: {
-      flexDirection: 'row',
-      gap: 3,
-      marginTop: 4,
-    },
-    calShiftDot: {
-      width: 6,
-      height: 6,
-      borderRadius: 3,
-    },
-    calLegend: {
-      flexDirection: 'row',
-      justifyContent: 'center',
-      gap: 16,
-      marginTop: 16,
-      paddingTop: 16,
-      borderTopWidth: 1,
-      borderTopColor: colors.border,
-      flexWrap: 'wrap',
-    },
-    calLegendItem: {
-      flexDirection: 'row',
-      alignItems: 'center',
-      gap: 6,
-    },
-    calLegendDot: {
-      width: 10,
-      height: 10,
-      borderRadius: 5,
-    },
-    calLegendText: {
-      fontSize: 12,
-      color: colors.muted,
-    },
-
-    modalOverlay: {
-      flex: 1,
-      backgroundColor: 'rgba(0, 0, 0, 0.7)',
-      justifyContent: 'center',
-      alignItems: 'center',
-    },
-    modalContent: {
-      backgroundColor: colors.card,
-      width: width - 48,
-      borderRadius: 16,
-      padding: 20,
-      borderWidth: 1,
-      borderColor: colors.border,
-    },
-    modalHeader: {
-      flexDirection: 'row',
-      justifyContent: 'space-between',
-      alignItems: 'center',
-      marginBottom: 20,
-    },
-    modalTitle: {
-      fontSize: 18,
-      fontWeight: '700',
-      color: colors.text,
-    },
-    modalCloseBtn: {
-      width: 28,
-      height: 28,
-      alignItems: 'center',
-      justifyContent: 'center',
-    },
-    modalCloseText: {
-      fontSize: 20,
-      color: colors.muted,
-    },
-    modalBody: {
-      gap: 12,
-    },
-    modalTitleRow: {
-      flexDirection: 'row',
-      justifyContent: 'space-between',
-      alignItems: 'center',
-      marginBottom: 12,
-    },
-    modalShiftTitle: {
-      fontSize: 16,
-      fontWeight: '700',
-      color: colors.text,
-      flex: 1,
-    },
-    statusBadge: {
-      paddingHorizontal: 12,
-      paddingVertical: 6,
-      borderRadius: 12,
-      marginLeft: 12,
-    },
-    statusBadgeText: {
-      fontSize: 12,
-      fontWeight: '600',
-      color: colors.white,
-    },
-    modalDetail: {
-      flexDirection: 'row',
-      justifyContent: 'space-between',
-    },
-    modalLabel: {
-      fontSize: 14,
-      color: colors.muted,
-    },
-    modalValue: {
-      fontSize: 14,
-      fontWeight: '500',
-      color: colors.text,
-      flexShrink: 1,
-      textAlign: 'right',
-    },
-    modalRequirements: {
-      marginTop: 12,
-      paddingTop: 16,
-      borderTopWidth: 1,
-      borderTopColor: colors.border,
-    },
-    modalRequirementsTitle: {
-      fontSize: 14,
-      fontWeight: '600',
-      color: colors.text,
-      marginBottom: 12,
-    },
-    modalTags: {
-      flexDirection: 'row',
-      gap: 8,
-      flexWrap: 'wrap',
-    },
-    modalTag: {
-      backgroundColor: colors.primarySoft,
-      paddingHorizontal: 12,
-      paddingVertical: 6,
-      borderRadius: 6,
-    },
-    modalTagText: {
-      fontSize: 12,
       color: colors.text,
     },
